Rename the Fixtures class in com.gu.support.workers to avoid collisions with the class of the same name in support-models
