+++ src/test/scala/com/gu/support/workers/WrapperSpec.scala
@@ -2,7 +2,7 @@ package com.gu.support.workers
 
 import com.gu.monitoring.SafeLogger
 import com.gu.support.encoding.CustomCodecs._
-import com.gu.support.workers.Fixtures.{contribution, wrapFixture}
+import com.gu.support.workers.JsonFixtures.{contribution, wrapFixture}
 import com.gu.support.workers.encoding.Conversions.StringInputStreamConversions
 import com.gu.support.workers.encoding.Encoding
 import com.gu.support.workers.states.CreateSalesforceContactState
@@ -18,7 +18,7 @@ class WrapperSpec extends FlatSpec with Matchers {
   }
 
   it should "be able to handle a JsonWrapper with messages" in {
-    val result = Encoding.in[CreateSalesforceContactState](Fixtures.wrapperWithMessages.asInputStream)
+    val result = Encoding.in[CreateSalesforceContactState](JsonFixtures.wrapperWithMessages.asInputStream)
     SafeLogger.info(s"$result")
     result.isSuccess should be(true)
   }
