+++ src/test/scala/com/gu/support/workers/errors/PayPalErrorsSpec.scala
@@ -6,7 +6,7 @@ import com.gu.config.Configuration
 import com.gu.okhttp.RequestRunners
 import com.gu.paypal.PayPalService
 import com.gu.support.config.PayPalConfig
-import com.gu.support.workers.Fixtures.{createPayPalPaymentMethodDigitalPackJson, wrapFixture}
+import com.gu.support.workers.JsonFixtures.{createPayPalPaymentMethodDigitalPackJson, wrapFixture}
 import com.gu.support.workers.LambdaSpec
 import com.gu.support.workers.exceptions.RetryUnlimited
 import com.gu.support.workers.lambdas.CreatePaymentMethod
