+++ src/test/scala/com/gu/support/workers/lambdas/CreateSalesforceContactDecoderSpec.scala
@@ -1,7 +1,7 @@
 package com.gu.support.workers.lambdas
 
 import com.gu.support.encoding.CustomCodecs._
-import com.gu.support.workers.Fixtures._
+import com.gu.support.workers.JsonFixtures._
 import com.gu.support.workers.states.CreateSalesforceContactState
 import com.gu.support.workers.{Contribution, PayPalReferenceTransaction, PaymentMethod}
 import com.typesafe.scalalogging.LazyLogging
